Correct data type for blurRadius

+++ RWBlurPopover/RWBlurPopoverView.h
@@ -12,7 +12,7 @@
 
 @interface RWBlurPopoverView : UIView
 
-@property (nonatomic, assign) int blurRadius;
+@property (nonatomic, assign) float blurRadius;
 @property (nonatomic, strong) UIView *blurView;
 @property (nonatomic, strong) UIView *container;
 @property (nonatomic, copy) dispatch_block_t dismissalBlock;
